Testing routes

+++ RSMSessionsEnterpriseIntegrations/Application/DTOs/SalesOrderHeader/UpdateSalesOrderHeaderDto.cs
@@ -11,15 +11,15 @@
         public bool OnlineOrderFlag { get; set; }
         public string? PurchaseOrderNumber { get; set; }
         public string? AccountNumber { get; set; }
-        public short CustomerId { get; set; }
-        public short SalesPersonID { get; set; }
-        public short? TerritoryID { get; set; }
-        public short BillToAddressID { get; set; }
-        public short ShipToAddressID { get; set; }
-        public short ShipMethodID { get; set; }
-        public short? CreditCardID { get; set; }
+        public int CustomerId { get; set; }
+        public int SalesPersonID { get; set; }
+        public int? TerritoryID { get; set; }
+        public int BillToAddressID { get; set; }
+        public int ShipToAddressID { get; set; }
+        public int ShipMethodID { get; set; }
+        public int? CreditCardID { get; set; }
         public string? CreditCardApprovalCode { get; set; }
-        public short? CurrencyRateID { get; set; }
+        public int? CurrencyRateID { get; set; }
         public decimal Subtotal { get; set; }
         public decimal TaxAmt { get; set; }
         public decimal Freight { get; set; }
